requirements.txt
APScheduler==3.5.1
asn1crypto==0.24.0
bcrypt==3.1.4
bleach==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2
dominate==2.3.1
Flask==0.12.2
Flask-APScheduler==1.8.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
multi-key-dict==2.0.3
paramiko==2.4.1
pbr==4.0.1
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.2
python-jenkins==0.4.16
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
tzlocal==1.5.1
urllib3==1.22
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
